chore: scope.js validates names itself
